Added tests for Modbus Connector
